Remove unnecessary Travis CI build steps
As of the previous commit, we're now resolving Bisq's custom bitcoinj
library via JitPack. This means it is no longer necessary in
before_script to clone, build and install bitcoinj at all.

The script: command has also been removed, as Travis' own defaults for
building Maven project are sufficient for our needs. We just need to
know if we've broken the build (and Slack notifications will be
configured for that in the subsequent commit).

See bitsquare/bitsquare#899